feat!: implement restartIfChanged option for containers
Previously, changing the container's system configuration would not
restart or reactivate container on activation. This change implements
a simple systemd restart trigger for the templated systemd-nspawn unit.

We also have to change the default value of systemd.nspawn(5)
KillSignal= in order to avoid SIGKILL-ing the container on restart and
causing potential data loss.

I think we could look into allowing activating new configurations right
in the container, to avoid a full restart, but for now this should be
good enough.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>